적과 기사의 공격 구현됨. * 죽음 및 게임 진행파트 개발필요

+++ 8.Battle/MonsterStateViewer.cs
@@ -31,7 +31,8 @@ public class MonsterStateViewer : MonoBehaviour
 
     public float GetHpWidthSize()
     {
-        float value = 100 * (bmp.ms.s.Hp/bmp.ms.m.hp);
+        float value = 100 * ((float)bmp.ms.s.Hp/(float)bmp.ms.m.hp);
+        
         return value;
     }
